requirements.txt
httpx
python-dotenv
fastapi
uvicorn
python-docx
docxtpl
python-multipart
pydantic
pandas
openpyxl
